fix: encode is128 flag for executors 01 and 02
