+++ src/config.ts
@@ -55,8 +55,8 @@ const baseConfigs: { [network: number]: BaseConfig } = {
     privateHttpProvider: process.env.HTTP_PROVIDER_1,
     augustusV6Address: '0x6a000f20005980200259b80c5102003040001068',
     executorsAddresses: {
-      Executor01: '0x000010036C0190E009a000d0fc3541100A07380A',
-      Executor02: '0x00C600b30fb0400701010F4b080409018B9006E0',
+      Executor01: '0x8faa0000c10015610005ca010ee000d006e0e820',
+      Executor02: '0x6f0538001f90d0a5f0000060d01d34c002030900',
       Executor03: '0xa000b020c290d000020aac04026b5306d60050f0',
     },
     adapterAddresses: {
@@ -154,8 +154,8 @@ const baseConfigs: { [network: number]: BaseConfig } = {
     bebopAuthName: process.env.API_KEY_BEBOP_AUTH_NAME || '',
     bebopAuthToken: process.env.API_KEY_BEBOP_AUTH_TOKEN || '',
     executorsAddresses: {
-      Executor01: '0x000010036C0190E009a000d0fc3541100A07380A',
-      Executor02: '0x00C600b30fb0400701010F4b080409018B9006E0',
+      Executor01: '0x8faa0000c10015610005ca010ee000d006e0e820',
+      Executor02: '0x6f0538001f90d0a5f0000060d01d34c002030900',
       Executor03: '0xa000b020c290d000020aac04026b5306d60050f0',
     },
     hashFlowAuthToken: process.env.API_KEY_HASHFLOW_AUTH_TOKEN || '',
@@ -191,8 +191,8 @@ const baseConfigs: { [network: number]: BaseConfig } = {
     privateHttpProvider: process.env.HTTP_PROVIDER_137,
     augustusV6Address: '0x6a000f20005980200259b80c5102003040001068',
     executorsAddresses: {
-      Executor01: '0x000010036C0190E009a000d0fc3541100A07380A',
-      Executor02: '0x00C600b30fb0400701010F4b080409018B9006E0',
+      Executor01: '0x8faa0000c10015610005ca010ee000d006e0e820',
+      Executor02: '0x6f0538001f90d0a5f0000060d01d34c002030900',
       Executor03: '0xa000b020c290d000020aac04026b5306d60050f0',
     },
     bebopAuthName: process.env.API_KEY_BEBOP_AUTH_NAME || '',
@@ -236,8 +236,8 @@ const baseConfigs: { [network: number]: BaseConfig } = {
     dexalotAuthToken: process.env.API_KEY_DEXALOT_AUTH_TOKEN || '',
     augustusV6Address: '0x6a000f20005980200259b80c5102003040001068',
     executorsAddresses: {
-      Executor01: '0x000010036C0190E009a000d0fc3541100A07380A',
-      Executor02: '0x00C600b30fb0400701010F4b080409018B9006E0',
+      Executor01: '0x8faa0000c10015610005ca010ee000d006e0e820',
+      Executor02: '0x6f0538001f90d0a5f0000060d01d34c002030900',
       Executor03: '0xa000b020c290d000020aac04026b5306d60050f0',
     },
     adapterAddresses: {
@@ -271,8 +271,8 @@ const baseConfigs: { [network: number]: BaseConfig } = {
       process.env[`HASHFLOW_DISABLED_MMS_146`]?.split(',') || [],
     augustusV6Address: '0x6a000f20005980200259b80c5102003040001068',
     executorsAddresses: {
-      Executor01: '0x67dd00d00d000003a383b096091f0a3060000d08',
-      Executor02: '0x700602C7b720200a09000f38B0d00Ee00c54f000',
+      Executor01: '0x8faa0000c10015610005ca010ee000d006e0e820',
+      Executor02: '0x6f0538001f90d0a5f0000060d01d34c002030900',
       Executor03: '0xa000b020c290d000020aac04026b5306d60050f0',
     },
     adapterAddresses: {},
@@ -303,8 +303,8 @@ const baseConfigs: { [network: number]: BaseConfig } = {
       process.env[`HASHFLOW_DISABLED_MMS_42161`]?.split(',') || [],
     augustusV6Address: '0x6a000f20005980200259b80c5102003040001068',
     executorsAddresses: {
-      Executor01: '0x000010036C0190E009a000d0fc3541100A07380A',
-      Executor02: '0x00C600b30fb0400701010F4b080409018B9006E0',
+      Executor01: '0x8faa0000c10015610005ca010ee000d006e0e820',
+      Executor02: '0x6f0538001f90d0a5f0000060d01d34c002030900',
       Executor03: '0xa000b020c290d000020aac04026b5306d60050f0',
     },
     dexalotAuthToken: process.env.API_KEY_DEXALOT_AUTH_TOKEN || '',
@@ -340,8 +340,8 @@ const baseConfigs: { [network: number]: BaseConfig } = {
     privateHttpProvider: process.env.HTTP_PROVIDER_10,
     augustusV6Address: '0x6a000f20005980200259b80c5102003040001068',
     executorsAddresses: {
-      Executor01: '0x000010036C0190E009a000d0fc3541100A07380A',
-      Executor02: '0x00C600b30fb0400701010F4b080409018B9006E0',
+      Executor01: '0x8faa0000c10015610005ca010ee000d006e0e820',
+      Executor02: '0x6f0538001f90d0a5f0000060d01d34c002030900',
       Executor03: '0xa000b020c290d000020aac04026b5306d60050f0',
     },
     hashFlowAuthToken: process.env.API_KEY_HASHFLOW_AUTH_TOKEN || '',
@@ -379,8 +379,8 @@ const baseConfigs: { [network: number]: BaseConfig } = {
     augustusRFQAddress: '0x92EaD5bACf6F0E995FA46Ad8215A9b11f67ca241',
     augustusV6Address: '0x6a000f20005980200259b80c5102003040001068',
     executorsAddresses: {
-      Executor01: '0x000010036c0190e009a000d0fc3541100a07380a',
-      Executor02: '0x00c600b30fb0400701010f4b080409018b9006e0',
+      Executor01: '0x8faa0000c10015610005ca010ee000d006e0e820',
+      Executor02: '0x6f0538001f90d0a5f0000060d01d34c002030900',
       Executor03: '0xa000b020c290d000020aac04026b5306d60050f0',
     },
     rpcPollingMaxAllowedStateDelayInBlocks: 0,
@@ -414,8 +414,8 @@ const baseConfigs: { [network: number]: BaseConfig } = {
     hashFlowDisabledMMs: [],
     augustusV6Address: '0x6a000f20005980200259b80c5102003040001068',
     executorsAddresses: {
-      Executor01: '0x000010036C0190E009a000d0fc3541100A07380A',
-      Executor02: '0x00C600b30fb0400701010F4b080409018B9006E0',
+      Executor01: '0x8faa0000c10015610005ca010ee000d006e0e820',
+      Executor02: '0x6f0538001f90d0a5f0000060d01d34c002030900',
       Executor03: '0xa000b020c290d000020aac04026b5306d60050f0',
     },
     adapterAddresses: {
@@ -447,8 +447,8 @@ const baseConfigs: { [network: number]: BaseConfig } = {
     augustusRFQAddress: '0x92EaD5bACf6F0E995FA46Ad8215A9b11f67ca241',
     augustusV6Address: '0x6a000f20005980200259b80c5102003040001068',
     executorsAddresses: {
-      Executor01: '0x000010036C0190E009a000d0fc3541100A07380A',
-      Executor02: '0x00C600b30fb0400701010F4b080409018B9006E0',
+      Executor01: '0x8faa0000c10015610005ca010ee000d006e0e820',
+      Executor02: '0x6f0538001f90d0a5f0000060d01d34c002030900',
       Executor03: '0xa000b020c290d000020aac04026b5306d60050f0',
     },
     rpcPollingMaxAllowedStateDelayInBlocks: 0,
@@ -478,8 +478,8 @@ const baseConfigs: { [network: number]: BaseConfig } = {
     rfqConfigs: {},
     hashFlowDisabledMMs: [],
     executorsAddresses: {
-      Executor01: '0x000010036c0190e009a000d0fc3541100a07380a',
-      Executor02: '0x00c600b30fb0400701010f4b080409018b9006e0',
+      Executor01: '0x8faa0000c10015610005ca010ee000d006e0e820',
+      Executor02: '0x6f0538001f90d0a5f0000060d01d34c002030900',
       Executor03: '0xa000b020c290d000020aac04026b5306d60050f0',
     },
     uniswapV2ExchangeRouterAddress:
@@ -505,8 +505,8 @@ const baseConfigs: { [network: number]: BaseConfig } = {
     augustusRFQAddress: '0x92EaD5bACf6F0E995FA46Ad8215A9b11f67ca241',
     augustusV6Address: '0x6a000f20005980200259b80c5102003040001068',
     executorsAddresses: {
-      Executor01: '0x000010036C0190E009a000d0fc3541100A07380A',
-      Executor02: '0x00C600b30fb0400701010F4b080409018B9006E0',
+      Executor01: '0x8faa0000c10015610005ca010ee000d006e0e820',
+      Executor02: '0x6f0538001f90d0a5f0000060d01d34c002030900',
       Executor03: '0xa000b020c290d000020aac04026b5306d60050f0',
     },
     rpcPollingMaxAllowedStateDelayInBlocks: 0,
