+++ src/executor/Executor03BytecodeBuilder.ts
@@ -1,4 +1,4 @@
-import { ethers, BigNumber } from 'ethers';
+import { ethers } from 'ethers';
 import { DexExchangeBuildParam } from '../types';
 import {
   Address,
@@ -19,11 +19,6 @@ import {
 const {
   utils: { hexlify, hexDataLength, hexConcat, hexZeroPad, solidityPack },
 } = ethers;
-
-/** Set bit 15 (uint128 write mode) on an Executor03 flag. */
-function applyIs128(flag: number): number {
-  return flag | 0x8000;
-}
 
 export type Executor03SingleSwapCallDataParams = {
   swap: OptimalSwap;
@@ -423,7 +418,7 @@ export class Executor03BytecodeBuilder extends ExecutorBytecodeBuilder<
       );
     }
 
-    const finalFlag = amountsPacked128 ? applyIs128(flag) : flag;
+    const finalFlag = amountsPacked128 ? this.applyIs128(flag) : flag;
 
     return this.buildCallData(
       exchangeParam.targetExchange,
@@ -503,94 +498,6 @@ export class Executor03BytecodeBuilder extends ExecutorBytecodeBuilder<
         swapsCalldata, // // calldata
       ],
     );
-  }
-
-  /**
-   * Find the position of an amount in calldata, trying both positive and
-   * negative encodings. For uint128 mode (is128), searches for 16-byte
-   * int128 patterns and returns the 32-byte slot position for mstore.
-   */
-  private findAmountPosWithFallback(
-    exchangeData: string,
-    amount: string,
-    is128: boolean,
-  ): number {
-    if (is128) {
-      return this.findAmount128PosInCalldata(exchangeData, amount);
-    }
-
-    // uint256 mode: try positive encoding first
-    const positiveEncoded = ethers.utils.defaultAbiCoder.encode(
-      ['uint256'],
-      [amount],
-    );
-    let pos = this.findAmountPosInCalldata(exchangeData, positiveEncoded);
-
-    // If not found, try negative int256 encoding
-    if (pos >= exchangeData.length / 2) {
-      const negativeEncoded = ethers.utils.defaultAbiCoder.encode(
-        ['int256'],
-        [BigNumber.from(amount).mul(-1)],
-      );
-      pos = this.findAmountPosInCalldata(exchangeData, negativeEncoded);
-    }
-
-    return pos;
-  }
-
-  /**
-   * Find the byte position of a 128-bit amount in calldata.
-   * Searches for both positive and negative int128 encodings.
-   * Returns the 32-byte slot position (16 bytes before the int128 value)
-   * so that mstore with uint128 mode writes to the correct lower 16 bytes.
-   */
-  private findAmount128PosInCalldata(
-    exchangeData: string,
-    amount: string,
-  ): number {
-    const rawCalldata = exchangeData.replace('0x', '');
-    const amountBN = BigNumber.from(amount);
-
-    // Try positive int128 encoding (32 hex chars = 16 bytes)
-    const positiveHex = hexZeroPad(
-      amountBN.toTwos(128).toHexString(),
-      16,
-    ).replace('0x', '');
-    let idx = this.findByteAligned(rawCalldata, positiveHex);
-
-    // Try negative int128 encoding if positive not found
-    if (idx === -1) {
-      const negativeHex = hexZeroPad(
-        amountBN.mul(-1).toTwos(128).toHexString(),
-        16,
-      ).replace('0x', '');
-      idx = this.findByteAligned(rawCalldata, negativeHex);
-    }
-
-    if (idx !== -1) {
-      // The int128 value starts at byte position idx/2.
-      // mstore with uint128 mode writes lower 16 bytes of a 32-byte slot,
-      // so the slot position is 16 bytes before the int128 value.
-      const slotPos = idx / 2 - 16;
-      if (slotPos >= 0) {
-        return slotPos;
-      }
-    }
-
-    // Not found — return past end of calldata (harmless write)
-    return exchangeData.length / 2;
-  }
-
-  /** Find a hex pattern in calldata, only accepting byte-aligned matches. */
-  private findByteAligned(rawCalldata: string, pattern: string): number {
-    for (
-      let idx = rawCalldata.indexOf(pattern);
-      idx !== -1;
-      idx = rawCalldata.indexOf(pattern, idx + 1)
-    ) {
-      if (idx % 2 === 0) return idx;
-    }
-    return -1;
   }
 
   private addMetadata(
